backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
psycopg[binary]>=3.2
gunicorn>=22.0
python-decouple>=3.8
sentry-sdk>=2.0

# Test
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3
faker>=26.0
